Add UseGoTimeValues Decoder option
When decoding to an interface{} value the current implementation of the
Decoder will always decode yaml values with the !!timestamp tag as
strings. According to the comments in the code this behavior ensures
backwards compatibility and will be dropped in V3 of the package.

This commit introduces a new Decoder option (useGoTimeValues) to deal
with cases where the package users expect the decoder to decode such
values into time.Time values. This option is false by default and can be
enabled using the SetUseGoTimeValues() method of the yaml decoder.